% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{calcAvgTable}
\alias{calcAvgTable}
\title{Calculate random selected background}
\usage{
calcAvgTable(coverage, name_table)
}
\arguments{
\item{coverage}{Coverage matrix (feature x cell)}

\item{name_table}{Table of feature names to include in background}
}
\value{
average coverage table for each feature and cell
}
\description{
Calculates the average coverage of each row of feature names in name_table
in order to produce a background average. Any averages of zero are replaced
with the smallest value for that cell.
}
