% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{calcCoverage}
\alias{calcCoverage}
\title{Calculate coverage}
\usage{
calcCoverage(data, df, avg_frag_length)
}
\arguments{
\item{data}{Matrix of copy numbers}

\item{df}{dataframe containing a feature lengths column}

\item{avg_frag_length}{Average length of a genomic framgnet}
}
\value{
coverage matrix
}
\description{
This fuction calculate coverage by multiplying read counts by the average
fragment length and dividing by the length of the feature
}
