% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{calcGC}
\alias{calcGC}
\title{Calculate GC content and append to dataframe}
\usage{
calcGC(df, tiles, GC_cutoff)
}
\arguments{
\item{df}{existing dataframe}

\item{tiles}{GRanges tiles object}

\item{GC_cutoff}{Maximum GC content to be included}
}
\value{
dataframe containing GC content, count, and frequency.
}
\description{
This function the GC content of GenomicRanges Tiles and appends it to
an existing dataframe. This function uses the Hsapiens genome from
BSgenome.Hsapiens.UCSC.hg19.
}
