% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{calcOverlap}
\alias{calcOverlap}
\title{Calculate overlap between feature bin and peaks}
\usage{
calcOverlap(df, tiles, peaks, genome, peak_occ_cutoff)
}
\arguments{
\item{df}{dataframe to contain the lengths without peaks}

\item{tiles}{GRanges tiles object}

\item{genome}{genetic sequence information}

\item{peak_off_cutoff}{Maximum percentage of peaks in a given feature}
}
\value{
dataframe containing the lengths of features without peaks and column of features above the peak cutoff
}
\description{
This function calculates the length of each Genomic feature not occupied by
peaks, and checks if the percentage of peaks is above a certain threshold.
The lengths and boolean values are appended as columns to the provided dataframe
}
