% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{GCplot}
\alias{GCplot}
\title{Plot Corrlation plot for potential bias factors}
\usage{
GCplot(fc, df, project, cormethod)
}
\arguments{
\item{fc}{Fold change matrix}

\item{df}{dataframe containing GC content data}

\item{project}{Name of the project}

\item{cormethod}{Method to use in the correlation calculation}
}
\value{
Creates a correlation plot and returns GC content
}
\description{
This function creates a correlation plot for potential bias factors using
GC content
}
