% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{getGCBackground}
\alias{getGCBackground}
\title{Select GC matched background}
\usage{
getGCBackground(coverage, df, GC_radius, GC_sample_size)
}
\arguments{
\item{coverage}{coverage matrix (as produced by calcCoverage())}

\item{df}{dataframe containing the GC content of genomic features}

\item{GC_radius}{half the number of neighbors that will be matched to each feature}

\item{GC_sample_size}{number of features that will be sampled to produce the background}
}
\value{
A matrix containing the names of GC matched features
}
\description{
This function matches each feature to a specified number of its closest
GC matched neighbors. A certain number of these neighbors are then
sampled to create a GC matched background
}
