% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{getOverlap}
\alias{getOverlap}
\title{Create offtarget and ontarget reads}
\usage{
getOverlap(fragments_file, peak_path, blacklist_path = NULL)
}
\arguments{
\item{fragments_file}{path to the fragments file}

\item{peak_path}{path to the peaks BED file}

\item{blacklist_path}{path to a blacklist BED file, if it exists}
}
\value{
A list containing the paths to the offtarget and ontarget files
}
\description{
This function uses the terminal to find overlap between fragments file and
peaks. Bedtools and Tabix need to be installed in order for this
preprocessing to occur
}
