% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{plotCNV}
\alias{plotCNV}
\title{Plot CNV heatmap}
\usage{
plotCNV(
  fc,
  k_means,
  k_means_cell_cut_off,
  gz,
  project,
  GC_Content = NULL,
  tad = NULL,
  target_t = NULL,
  hc_cluster = 6
)
}
\arguments{
\item{fc}{Fold change matrix}

\item{k_means}{Number of desired clusters k-means (hiearchical if 0)}

\item{k_means_cell_cut_off}{minimum number of cells for a cluster to be seen}

\item{gz}{chromosome length data, produced by seqlengths()}

\item{project}{Name of project}

\item{GC_Content}{GC content data. See GCplot}

\item{tad}{tad data. See TADplot}

\item{target_t}{On-target content data. See ATACplot}

\item{hc_cluster}{number of hiearchical clusters}
}
\value{
creates heatmap and returns corresponding dataframe
}
\description{
This function creates a heatmap using CNV data, where each row is a cell
and reach column is a genomic feature. Cells can be clustered together
hierarchically (if k_means ==0), or with k-means clustering
}
