% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{removeUnmapped}
\alias{removeUnmapped}
\title{Remove unmappable regions from genomic tiles}
\usage{
removeUnmapped(tiles, unmapped_region)
}
\arguments{
\item{tiles}{GRanges tiles object}

\item{unmapped_region}{path to a .csv defining the unmappable regions}
}
\value{
GRanges object with the unmappable regions removed
}
\description{
This function removes any unmapped region defined by a .csv
from a GenomicRanges tile object
}
