% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{superCells}
\alias{superCells}
\title{Supercell - PCA & hierarchical clustering}
\usage{
superCells(fc, cellsPergroup)
}
\arguments{
\item{fc}{Fold change matrix}

\item{cellsPerGroup}{number of cells per supercell}
}
\value{
dataframe containing the supercell groups
}
\description{
Clusters cells into super cells using Principle Component Analysis
}
