% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{superCells_merge}
\alias{superCells_merge}
\title{Supercell - merge cells}
\usage{
superCells_merge(data, supercells)
}
\arguments{
\item{data}{matrix containing cnv data}

\item{supercells}{matrix containing supercell grouping assignemnts}
}
\value{
supercells with calculated cnv data
}
\description{
This function merges the cnv data of a group of cells according
to their supercell grouping
}
