% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{TADplot}
\alias{TADplot}
\title{Plot Corrlation plot for potential bias factors using TAD data}
\usage{
TADplot(fc, tiles, df, project, TAD_cluster)
}
\arguments{
\item{fc}{Fold change matrix}

\item{tiles}{GRanges tiles object}

\item{df}{dataframe}

\item{project}{Name of the project}

\item{TAD_Cluster}{file path to TAD cluster file}
}
\value{
Creates a correlation plot and returns TAD data
}
\description{
This function creates a correlation plot for potential bias factors using
TAD cluster data
}
