% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions.R
\name{tileTodf}
\alias{tileTodf}
\title{Convert Genomic Tiles to a dataframe}
\usage{
tileTodf(tiles)
}
\arguments{
\item{tiles}{GRanges Genomic Tiles Object}
}
\value{
A dataframe with genomic features as row names
}
\description{
This function converts a GRanges object to a dataframe
with appropriate row names
}
